feat: 添加DeepSpeed支持并优化训练配置
- 添加DeepSpeed Zero2/Zero3配置文件
- 新增accelerate训练脚本支持多节点分布式训练
- 优化数据类型转换逻辑，添加dtype_utils工具模块
- 更新依赖项，添加accelerate和deepspeed包
- 修复MKL多线程问题，添加环境变量设置
- 支持FP16/BF16/FP32混合精度训练
- 优化模型初始化兼容DeepSpeed Zero3
- 添加行为数据集路径环境变量覆盖功能
- 修复softmax数值稳定性问题
- 更新训练配置参数，支持梯度累积

+++ src/openpi/models_pytorch/pi05_subtask.py
@@ -307,7 +307,7 @@ class PI05SubtaskPytorch(PI0Pytorch):
 
         for step in range(max_tokens):
             if temperature > 0:
-                probs = torch.softmax(logits[:, -1] / temperature, dim=-1)
+                probs = torch.softmax(logits[:, -1] / temperature, dim=-1, dtype=torch.float32).to(logits.dtype)
                 next_token = torch.multinomial(probs, 1)
             else:
                 next_token = logits[:, -1].argmax(dim=-1, keepdim=True)
